(feat:JS): export 5 functions
